src/ui/custom_components/codeanalysisviewer/demo/requirements.txt
gradio_codeanalysisviewer